requirements.txt
dash==2.9.3
dash-ag-grid==2.0.0rc2
dash-bootstrap-components==1.2.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Session==0.4.0
pandas==1.5.1
plotly==5.13.0
python-dotenv==0.19.1
requests==2.28.2
dash-extensions==1.0.1
dash-iconify==0.1.2
gunicorn==20.1.0
werkzeug==2.2.3
stravalib=1.3.3
polyline==2.0.2